refactor(cli): flatten resource command module
Move the Resource command slice to the neutral command root and rename its internal wrapper names without changing public CLI or archive behavior.
